==> code/10-AnnotatedORFs.Rmd <==
---
title: "8-Annotated ORFs"
author: "Kathryn Walters"
date: "2023-05-16"
output: html_document
editor_options: 
  chunk_output_type: console
---


```{r libraries, warning=FALSE, message=FALSE, echo=FALSE}
library(Biostrings)
library(rlist)
library(ggVennDiagram)
library(readr)
library(here)
library(dplyr)
library(tidyr)
library(ggplot2)
library(ggthemes)

```


```{r}
## Summary of RMD contents
#In this file I am looking at both RiboCode and ORFquant software's ability to called canonical annotated ORFs. Then I want to see if those ORFs are identified as proteins in the associated proteomics data from sBCs. The data here is used in Figure 2. 

#1. Loading Data

#2. Identifying annotated ORFs with proteomic support in sBCs

#3. Using VennDiagram package to determine overlap in annotated ORFs called
```


# 1. Loading Data


```{r load data, message=FALSE, warning=FALSE, echo=FALSE}
AllORFs <- read_csv(here::here("output", "GeneCollapsedORFs_AllPrograms.csv"))
collapsedORFlist <- read_csv(here::here("output", "final_collapsedORFsList.csv"))


#getting RiboCode list of ORFs found in sBC proteomics
proteomicsRC <- read_tsv(here("proteomics", "RiboCode", "sBC_nORF_Ribocode", "combined_protein.tsv")) %>% filter(`Combined Unique Spectral Count` > 1) %>% pull(`Protein ID`) %>% as.data.frame()
colnames(proteomicsRC) <- "ORF_ID"

#getting ORFquant list of ORFs found in proteomics
proteomicsOQ <- read_tsv(here("proteomics", "ORFQuant", "sBC_nORF_ORFquant", "combined_protein.tsv")) %>% filter(`Combined Unique Spectral Count` > 1) %>% pull(`Protein ID`) %>% as.data.frame()
colnames(proteomicsOQ) <- "ORF_ID"

geneInfo <- read_csv(here("accessories","gencode.v26.primary.info.csv.zip"), col_names = F, show_col_types = FALSE) 
colnames(geneInfo) <- c("gene_id","transcript_id","biotype","symbol")

tx2gene <- geneInfo[,c(2,1)]
```


# 2. Identifying annotated ORFs with proteomic support in sBCs


```{r per software, message=FALSE, warning=FALSE, echo=FALSE}

#grabbing only annotated genes
RC_anno <- AllORFs %>% filter(ORF_type == "annotated" & program == "Ribocode")
OQ_anno <- AllORFs %>% filter(ORF_type == "annotated" & program == "ORFquant")
all_anno <- collapsedORFlist %>% filter(ORF_type == "annotated")

#asking if these identified ORFs have proteomic support
RC_anno <- RC_anno %>% dplyr::mutate(protein = case_when(ORF_ID %in% proteomicsRC$ORF_ID ~ 'Yes', !(ORF_ID %in% proteomicsRC$ORF_ID) ~ 'No'))
RC_anno$protein %>% table()

OQ_anno <- OQ_anno %>% dplyr::mutate(protein = case_when(ORF_ID %in% proteomicsOQ$ORF_ID ~ 'Yes', !(ORF_ID %in% proteomicsOQ$ORF_ID) ~ 'No'))
OQ_anno$protein %>% table()


#making a table to get the number/count of ORF ids with proteomic support and those without. 
t_RC <- RC_anno$protein %>% table() %>% as.data.frame()
t_RC$program <- "Ribocode"
colnames(t_RC) <- c("proteinSupport", "value", "program")

t_OQ <- OQ_anno$protein %>% table() %>% as.data.frame()
t_OQ$program <- "ORFquant"
colnames(t_OQ) <- c("proteinSupport", "value", "program")

barplotData <- rbind(t_RC, t_OQ)

#making plots of combined software ORFs
p_anno_proteoSupport_percet <- ggplot(barplotData, aes(fill=proteinSupport, y=value, x=program)) + 
    geom_bar(position="fill", stat="identity") + theme_minimal() + scale_fill_colorblind() + ylab("Percent") + xlab("")

p_anno_proteoSupport_percet


p_anno_proteoSupport_count <- ggplot(barplotData, aes(fill=proteinSupport, y=value, x=program)) + 
    geom_bar(position="stack", stat="identity") + theme_minimal() + scale_fill_colorblind() + ylab("Counts") + xlab("")

p_anno_proteoSupport_count

```


# 3. Using VennDiagram to determine overlap in annotated ORFs called


```{r Overlap between programs, message=FALSE, warning=FALSE}
#grabbing geneIDs associated with each program's already filtered annotated genes.
set1 <- as.matrix(RC_anno)[,2]
set3 <- as.matrix(OQ_anno)[,2]

listALL <- list(RiboCode=set1, ORFquant = set3)

#using VennDiagram package to get overlap of these gene lists. 
unique.partition <- VennDiagram::get.venn.partitions(listALL)
colnames(unique.partition) <- c("Ribocode", "ORFquant", "Set", "geneID", "count")

#now creating separate lists of IDs that are shared or unique to each dataset. 
combined <- unique.partition[1,4] %>% unlist()
riboOnly <- unique.partition[3,4] %>% unlist()
ORFquantOnly <- unique.partition[2,4] %>% unlist()

#for each list, do they have proteomic support?
proteoList <- rbind(RC_anno[,c(2,6)], OQ_anno[,c(2,6)]) %>% unique()

#if that sequence ever had protein support, it is counted as protein support.
proteoList <- proteoList %>%
  group_by(AAseq) %>%
  summarize(protein = if_else(any(protein == "Yes"), "Yes", "No"), .groups = "drop")

#are they found in one dataset or both?
proteoList <- proteoList %>% dplyr::mutate(Venndiagram = case_when(AAseq %in% combined ~ 'Both', AAseq %in% riboOnly ~ 'Ribocode', AAseq %in% ORFquantOnly ~ "ORFquant"))

#formatting to make a bar plot.
barplotInfoVenn <- proteoList %>% group_by(Venndiagram) %>% dplyr::select(protein) %>% table() %>% as.data.frame()

#creating a barplot of the percent frequency of a gene having proteomic support if it was called by one software or both.
p_anno_proteoSupport_percent_overlap <- ggplot(barplotInfoVenn, aes(fill=protein, y=Freq, x=Venndiagram)) + geom_bar(position="fill", stat="identity") + theme_minimal() + scale_fill_colorblind() + ylab("Percent") + xlab("")


#ggsave(plot = p_anno_proteoSupport_percent_overlap, filename = here("plots", "Figure2", "Fig2F_annoProteinSupport_percent.pdf"), width = 4, height = 3)


#can look at absolute counts also. However, this data is better viewed as venn diagram I think so this bar graph was not included in figures. 
p_anno_proteoSupport_count_Venn <- ggplot(barplotInfoVenn, aes(fill=protein, y=Freq, x=Venndiagram)) + geom_bar(position="stack", stat="identity") + theme_minimal() + scale_fill_grey() + ylab("Counts") + xlab("")

p_anno_proteoSupport_count_Venn


ggVenn <- ggVennDiagram(listALL, edge_color = "black", label = "count", label_alpha = 0) + theme(legend.position = "none")

ggVenn

ggsave(plot = ggVenn, filename = here("plots","Figure4", "Fig4B_annoORFs_Venndiagram.pdf"), width = 4, height = 3)

```
