==> code/16-Islet_proteomics.Rmd <==
---
title: "10-Islet_proteomics"
author: "Kathryn Walters"
date: "2023-07-07"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r libraries, warning=FALSE, message=FALSE, echo=FALSE}
library(Biostrings)
library(rlist)
library(ggVennDiagram)
library(readr)
library(here)
library(dplyr)
library(tidyr)
library(ggplot2)
library(ggthemes)
library(limma)
library(tidyverse)
library(readxl)
```


```{r}
## Summary of RMD contents
#In this file I am determining if the anno ORFs and nuORFs we identified are also found in Human islet proteomics data. The plots here are used in Figure 4 and Supplementary Figure 4. 

#1. Loading Data

#2. Loading proteomics data

#3. Proteomic Support for Annotated genes in Human Islets

#4. Proteomic Support for nuORFs in Human Islets
  #a. Looking at all nuORFs - not separated into categories
  #b. Looking at nuORF categories separated by programs
  #c. Final figures - nuORF categories with program collapsed. 
```


# 1. Loading Data


```{r load data, message=FALSE, warning=FALSE, echo=FALSE}
AllORFs <- read_xlsx(here("output", "AllORFs_wProteomics_filtered_plus.xlsx"))
AllORFs_NP <- read_xlsx(here("output", "AllORFs_wProteomics_filtered_ProgramCollapsed_plus.xlsx"))


geneInfo <- read_csv(here("accessories","gencode.v26.primary.info.csv.zip"), col_names = F, show_col_types = FALSE) 
colnames(geneInfo) <- c("gene_id","transcript_id","biotype","symbol")

tx2gene <- geneInfo[,c(2,1)]
```


# 2. Loading proteomics data


```{r setup, include=FALSE, message=FALSE, warning=FALSE}
#getting ORFquant list of ORFs found in proteomics
islets2_OQ <- read_tsv(here("proteomics", "ORFQuant", "Cyto24_set2", "ORF_Cytro24_set2_protein.tsv")) %>% filter(`Unique Spectral Count` > 1) %>% pull(`Protein ID`)%>% as.data.frame()
colnames(islets2_OQ) <- "ORF_ID"

islets3_OQ <- read_tsv(here("proteomics", "ORFQuant", "Cyto24_set3", "ORF_Cyto24_set3_protein.tsv")) %>% filter(`Unique Spectral Count` > 1) %>% pull(`Protein ID`)%>% as.data.frame()
colnames(islets3_OQ) <- "ORF_ID"

islets_OQall <- rbind(islets2_OQ, islets3_OQ) %>% unique()


#getting RiboCode list of ORFs found in islet proteomics
islets2_RC <- read_tsv(here("proteomics", "RiboCode", "Cyto24_set2", "Ribo_Cyto24_set2_protein.tsv")) %>% filter(`Unique Spectral Count` > 1) %>% pull(`Protein ID`)%>% as.data.frame()
colnames(islets2_RC) <- "ORF_ID"

islets3_RC <- read_tsv(here("proteomics", "RiboCode", "Cyto24_set3", "Ribo_Cyto24_set3_protein.tsv")) %>% filter(`Unique Spectral Count` > 1) %>% pull(`Protein ID`)%>% as.data.frame()
colnames(islets3_RC) <- "ORF_ID"


islets_RCall <- rbind(islets2_RC, islets3_RC) %>% unique()
isletsALL <- rbind(islets_RCall, islets_OQall)
```


# 3. Proteomic Support for Annotated ORFs in Human Islets


```{r per software, message=FALSE, warning=FALSE, echo=FALSE}
#grabbing only annotated genes
RC_anno <- AllORFs %>% filter(ORF_type == "annotated" & program == "Ribocode")
OQ_anno <- AllORFs %>% filter(ORF_type == "annotated" & program == "ORFquant")

#asking if these identified ORFs have proteomic support
#Start with ribocode 
RC_anno <- RC_anno %>% dplyr::mutate(islet = case_when(ORF_ID %in% islets_RCall$ORF_ID ~ 'Yes', !(ORF_ID %in% islets_RCall$ORF_ID) ~ 'No'))
RC_anno$islet %>% table()


#finally ORFquant
OQ_anno <- OQ_anno %>% dplyr::mutate(islet = case_when(ORF_ID %in% islets_OQall$ORF_ID ~ 'Yes', !(ORF_ID %in% islets_OQall$ORF_ID) ~ 'No'))
OQ_anno$islet %>% table()

#making a table to get the number/count of ORF ids with proteomic support and those without. 

# compute unique levels in data frame
lvls <- unique(unlist(RC_anno[20]))
  
# apply the summation per value for Ribocode
t_RC <- sapply(RC_anno[20], 
               function(x) table(factor(x, levels = lvls, 
                                        ordered = TRUE))) %>% as.data.frame()

t_RC$program <- "Ribocode"
t_RC <- t_RC %>% rownames_to_column("ProteinSupport") %>% pivot_longer(!c(program, ProteinSupport), names_to = "cell_type", values_to = "count")


# apply the summation per value for ORFquant
t_OQ <- sapply(OQ_anno[20], 
               function(x) table(factor(x, levels = lvls, 
                                        ordered = TRUE))) %>% as.data.frame()
t_OQ$program <- "ORFquant"
t_OQ <- t_OQ %>% rownames_to_column("ProteinSupport") %>% pivot_longer(!c(program, ProteinSupport), names_to = "cell_type", values_to = "count")


barplotData <- rbind(t_RC, t_OQ)


#making plots
p_anno_proteoSupport_percet_cellLines <- ggplot(barplotData, aes(fill=ProteinSupport, y=count, x=cell_type)) + 
    geom_bar(position="fill", stat="identity") +
  facet_wrap(~program) + theme_minimal() + scale_fill_colorblind() + ylab("Percent") + xlab("")

p_anno_proteoSupport_percet_cellLines


p_anno_proteoSupport_count_cellLines <- ggplot(barplotData, aes(fill=ProteinSupport, y=count, x=cell_type)) + 
    geom_bar(position="stack", stat="identity") +
  facet_wrap(~program) + theme_minimal() + scale_fill_colorblind() + ylab("Counts") + xlab("")

p_anno_proteoSupport_count_cellLines

```


# 4. Proteomic Support for nuORFs in Human Islets


```{r proteomic support all nuORFs, message=FALSE, warning=FALSE, echo=FALSE}
#First will look at nuORFs overall - not separated into categories. 

#grabbing only NON-annotated genes
NOTannoORFs <- AllORFs %>% filter(ORF_type != "annotated")

RC_NotAnno <- NOTannoORFs %>% filter(program == "Ribocode")
OQ_NotAnno <- NOTannoORFs %>% filter(program == "ORFquant")


#asking if these identified nuORFs have proteomic support
#Start with ribocode 
RC_NotAnno <- NOTannoORFs %>% dplyr::mutate(islet = case_when(ORF_ID %in% islets_RCall$ORF_ID ~ 'Yes', !(ORF_ID %in% islets_RCall$ORF_ID) ~ 'No'))
RC_NotAnno$islet %>% table()


#finally ORFquant
OQ_NotAnno <- NOTannoORFs %>% dplyr::mutate(islet = case_when(ORF_ID %in% islets_OQall$ORF_ID ~ 'Yes', !(ORF_ID %in% islets_OQall$ORF_ID) ~ 'No'))
OQ_NotAnno$islet %>% table()


#making a table to get the number/count of ORF ids with proteomic support and those without. 
  
# apply the summation per value for Ribocode
t_RC_NA <- sapply(RC_NotAnno[20], 
               function(x) table(factor(x, levels = lvls, 
                                        ordered = TRUE))) %>% as.data.frame()

t_RC_NA$program <- "Ribocode"
t_RC_NA <- t_RC_NA %>% rownames_to_column("ProteinSupport") %>% pivot_longer(!c(program, ProteinSupport), names_to = "cell_type", values_to = "count")


# apply the summation per value for ORFquant
t_OQ_NA <- sapply(OQ_NotAnno[20], 
               function(x) table(factor(x, levels = lvls, 
                                        ordered = TRUE))) %>% as.data.frame()
t_OQ_NA$program <- "ORFquant"
t_OQ_NA <- t_OQ_NA %>% rownames_to_column("ProteinSupport") %>% pivot_longer(!c(program, ProteinSupport), names_to = "cell_type", values_to = "count")


barplotData_NotAnno <- rbind(t_RC_NA, t_OQ_NA)

#making plots
p_NonAnno_proteoSupport_percet_cellLines <- ggplot(barplotData_NotAnno, aes(fill=ProteinSupport, y=count, x=cell_type)) + 
    geom_bar(position="fill", stat="identity") +
  facet_wrap(~program) + theme_minimal() + scale_fill_colorblind() + ylab("Percent") + xlab("")

p_NonAnno_proteoSupport_percet_cellLines


p_NonAnno_proteoSupport_count_cellLines <- ggplot(barplotData_NotAnno, aes(fill=ProteinSupport, y=count, x=cell_type)) + 
    geom_bar(position="stack", stat="identity") +
  facet_wrap(~program) + theme_minimal() + scale_fill_colorblind() + ylab("Counts") + xlab("")

p_NonAnno_proteoSupport_count_cellLines
```


```{r proteomic support nuORF categories per program, message=FALSE, warning=FALSE, echo=FALSE}
#Now going to do the same as above but broken into the individual nuORFs. Use facet_grid!

proteomicsIslet <- rbind(islets_RCall, islets_OQall)

AllORFs <- AllORFs %>% dplyr::mutate(islet = case_when(ORF_ID %in% proteomicsIslet$ORF_ID ~ 'Yes', !(ORF_ID %in% proteomicsIslet$ORF_ID) ~ 'No')) 


AllORFs_proteo <- AllORFs[,c(2,4:5)]

islet_freq <- AllORFs_proteo %>% group_by(ORF_type, program, islet) %>% dplyr::summarise(Freq = n())
colnames(islet_freq) <- c("ORF_type", "program", "support", "freq")
islet_freq$cell <- "islet"

tableProteomicSupport <- islet_freq


tableProteomicSupport$ORF_type <- factor(tableProteomicSupport$ORF_type, order = TRUE, levels =c("annotated", 'novel', 'uORF', "dORF", "other"))
levels(tableProteomicSupport$ORF_type)

colors <- c("#969696", "#009E73", "#56B4E9", "#E69F00", "#CC79A7")

p_nuORF_proteoSupport_percent <- ggplot(tableProteomicSupport %>% filter(cell != "sBC" & ORF_type != "other"), aes(fill=support, y=freq, x=cell)) + geom_bar(position="fill", stat="identity") + theme_minimal() + scale_fill_manual(values = colors) + ylab("Percent Protein Support") + xlab("") +
  facet_grid(program ~ ORF_type) + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

p_nuORF_proteoSupport_percent

p_nuORF_proteoSupport_count <- ggplot(tableProteomicSupport %>% filter(ORF_type != "annotated" & cell != "sBC"), aes(fill=support, y=freq, x=cell)) + geom_bar(position="stack", stat="identity") + theme_minimal() + scale_fill_colorblind() + ylab("Count of ORFs w/Proteomic Support") + xlab("") +
  facet_grid(program~ ORF_type) + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

p_nuORF_proteoSupport_count
```


```{r proteomic support nuORF categories, warning=FALSE, message=FALSE, echo=FALSE}
#now doing the same as above, looking at specific nuORFs, but don't have it faceted by program. 

AllORFs_collapsed <- AllORFs %>% column_to_rownames("ORF_ID")
AllORFs_collapsed <- AllORFs_collapsed[,-c(3)] %>% unique()

AllORFs_proteo_nProgram <- AllORFs[,c(2,4:5)]

islet_freq_NP <- AllORFs_proteo_nProgram %>% group_by(ORF_type, islet) %>% dplyr::summarise(Freq = n())
colnames(islet_freq_NP) <- c("ORF_type", "support", "freq")
islet_freq_NP$cell <- "islet"


tableProteomicSupport_NP <- islet_freq_NP


tableProteomicSupport_NP$ORF_type <- factor(tableProteomicSupport_NP$ORF_type, order = TRUE, levels =c("annotated", 'novel', 'uORF', "dORF", "other"))
levels(tableProteomicSupport_NP$ORF_type)


p_nuORF_proteoSupport_percent_NP <- ggplot(tableProteomicSupport_NP %>% filter(cell == "islet" & ORF_type != "other"), aes(fill=support, y=freq, x=ORF_type)) + geom_bar(position="fill", stat="identity") + theme_minimal() + scale_fill_colorblind() + ylab("Percent Protein Support") + xlab("") + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

p_nuORF_proteoSupport_percent_NP

#ggsave(plot = p_nuORF_proteoSupport_percent_NP, filename = here("plots", "Figure4", "Fig4A_nuORFProteinSupport_percent_cells_ORFfacet_NP_islet.pdf"), width = 4, height = 3)


p_nuORF_proteoSupport_count_NP <- ggplot(tableProteomicSupport_NP %>% filter(cell == "islet" & ORF_type =="annotated"), aes(fill=support, y=freq, x=cell)) + geom_bar(position="stack", stat="identity") + theme_minimal() + scale_fill_colorblind() + ylab("Count of ORFs w/Proteomic Support") + xlab("") +
  facet_grid(~ORF_type) + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

p_nuORF_proteoSupport_count_NP

#ggsave(plot = p_nuORF_proteoSupport_count_NP, filename = here("plots", "Figure4", "Supp4A_nuORFProteinSupport_counts_cells_ORFfacet_NP.pdf"), width = 7, height = 3)

p_nuORF_proteoSupport_count_NP2 <- ggplot(tableProteomicSupport_NP %>% filter(cell == "islet" & ORF_type !="annotated" & ORF_type !="other"), aes(fill=support, y=freq, x=cell)) + geom_bar(position="stack", stat="identity") + theme_minimal() + scale_fill_colorblind() + ylab("Count of ORFs w/Proteomic Support") + xlab("") +
  facet_grid(~ORF_type) + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

p_nuORF_proteoSupport_count_NP2

#ggsave(plot = p_nuORF_proteoSupport_count_NP2, filename = here("plots", "Figure4", "Supp4A_nuORFProteinSupport_counts_cells_ORFfacet_NP2.pdf"), width = 7, height = 3)

```


```{r}

isletsYes <- AllORFs %>% select(c(islet, ORF_ID))

ORFlist <- left_join(ORFlist, isletsYes)

ORFlist %>% filter(ORF_type != "annotated") %>% select(islet, protein) %>% table()
#so 283 of them are not detected by proteomics in sBCs or islets. 
#50 are detected in sBCs but not islets
#190 in islets but not sBCs
#566 in both


ORFlist %>% filter(ORF_type != "annotated") %>% select(islet, protein, HLA_presented_hi) %>% table()
#29 are detected at all levels -> islets, immuno, sBCs
#4 are detected by immuno only

ORFlist %>% filter(ORF_type != "annotated") %>% select(islet, protein, HLA_presented_lo) %>% table()
#183 are detected at all levels -> islets, immuno, sBCs
#82 are detected by immuno only but these could be overlap with anno proteins, and indeed is likely I think. However, from a database building standpoint, we could include them anyway. 

ImmunoLo <- read_csv(here("output", "HLA_presented_nuORF_lo.csv"))

ORFlist <- left_join(ORFlist, ImmunoLo %>% select(ORF_ID, HLA_presented_hi, HLA_presented_lo))
```



```{r}
isletTable <- AllORFs_NP %>% select(islet, ORF_type) %>% table() %>% as.data.frame()

isletTable <- isletTable %>%
  group_by(ORF_type) %>%
  mutate(Percentage = (Freq / sum(Freq)) * 100)

isletTable$cell <- "islet"
colnames(isletTable) <- c("proteinSupport", "ORF_type", "Freq","Percent", "Cell")




sBCTable <- AllORFs_NP %>% select(protein, ORF_type) %>% table() %>% as.data.frame()
sBCTable <- sBCTable %>%
  group_by(ORF_type) %>%
  mutate(Percentage = (Freq / sum(Freq)) * 100)

sBCTable$cell <- "sBC"
colnames(sBCTable) <- c("proteinSupport", "ORF_type", "Freq", "Percent", "Cell")

proteomALL <- rbind(isletTable, sBCTable)

proteomALL$ORF_type <- factor(proteomALL$ORF_type, order = TRUE, levels =c("annotated", 'novel', 'uORF', "dORF", "other"))
levels(proteomALL$ORF_type)

colors <- c("#969696", "#009E73", "#56B4E9", "#E69F00", "#CC79A7")


proteoALL_bar <- ggplot(proteomALL %>% filter(ORF_type != "other" & proteinSupport == "Yes"), aes(fill=ORF_type, y=Percent, x=Cell)) + geom_bar(position="dodge", stat="identity") + theme_minimal() + scale_fill_manual(values = colors) + ylab("Percent Protein Support") + xlab("") + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))


#ggsave(plot = proteoALL_bar, filename = here("plots", "Fig4_proteoALL_bar.pdf"), width = 5, height = 3)


proteoALL_bar_counts <- ggplot(proteomALL %>% filter(ORF_type != "other"), aes(fill=proteinSupport, y=Freq, x=ORF_type)) + geom_bar(position="stack", stat="identity") + theme_minimal() + scale_fill_manual(values = colors) + ylab("Percent Protein Support") + xlab("") + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) + facet_wrap(~Cell)

p1 <- ggplot(proteomALL %>% 
         filter(ORF_type != "other" & Cell == "sBC") %>% 
         mutate(ORF_group = ifelse(ORF_type == "annotated", "Annotated", "Other ORFs")),
       aes(fill = proteinSupport, y = Freq, x = ORF_type)) +
  geom_bar(position = "stack", stat = "identity") +
  theme_minimal() +
  scale_fill_manual(values = colors) +
  ylab("Percent Protein Support") +
  xlab("") + labs(title = "sBC") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1)) +
  facet_wrap(~ORF_group, scales = "free", ncol = 2)

p2 <- ggplot(proteomALL %>% 
         filter(ORF_type != "other" & Cell == "islet") %>% 
         mutate(ORF_group = ifelse(ORF_type == "annotated", "Annotated", "Other ORFs")),
       aes(fill = proteinSupport, y = Freq, x = ORF_type)) +
  geom_bar(position = "stack", stat = "identity") +
  theme_minimal() +
  scale_fill_manual(values = colors) +
  ylab("Percent Protein Support") +
  xlab("") + labs(title = "islet") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1)) +
  facet_wrap(~ORF_group, scales = "free", ncol = 2)

proteoCounts <- plot_grid(p1, p2)


#ggsave(plot = proteoCounts, filename = here("plots", "SuppFig4F_proteoALLcounts_bar.pdf"), width = 10, height = 3)
```



```{r}
AllORFs <- AllORFs %>% dplyr::mutate(islet = case_when(ORF_ID %in% isletsALL$ORF_ID ~ 'Yes', !(ORF_ID %in% isletsALL$ORF_ID) ~ 'No'))

AllORFs_NP <- AllORFs_NP %>% dplyr::mutate(islet = case_when(ORF_ID %in% isletsALL$ORF_ID ~ 'Yes', !(ORF_ID %in% isletsALL$ORF_ID) ~ 'No'))

AllORFs <- left_join(AllORFs, geneInfo[,c(1,4)] %>% unique())
AllORFs_NP <- left_join(AllORFs_NP, geneInfo[,c(1,4)] %>% unique())

AllORFs %>% filter(ORF_type != "annotated") %>% dplyr::select(islet, program) %>% table()

writexl::write_xlsx(AllORFs, here("output", "AllORFs_wProteomics_filtered_plus.xlsx"))
writexl::write_xlsx(AllORFs_NP, here("output", "AllORFs_wProteomics_filtered_ProgramCollapsed_plus.xlsx"))

```
